requirements.txt
pandas==2.0.3
easyocr==1.7.1
numpy==1.24.4
opencv-python==4.9.0
Pillow==10.1.0